requirements.txt
aniso8601==9.0.1
charset-normalizer==2.0.3
certifi==2021.5.30
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
aniso8601==9.0.1
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
webdriver-manager==3.4.2
Werkzeug==2.0.1
pymongo==3.12.0
aniso8601==9.0.1
charset-normalizer==2.0.3
certifi==2021.5.30
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1

google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gspread==4.0.1
gspread-dataframe==3.2.1
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.3.0

aniso8601==9.0.1
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
webdriver-manager==3.4.2
Werkzeug==2.0.1
pymongo==3.12.0
aniso8601==9.0.1
charset-normalizer==2.0.3
certifi==2021.5.30
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1

google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gspread==4.0.1
gspread-dataframe==3.2.1
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.3.0